Move server initialization code into separate package
